requirement.txt
aniso8601==9.0.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Faker==25.8.0
flask==3.0.3
Flask-Bcrypt==1.0.1
flask-marshmallow==1.2.1
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mysqlclient==2.2.4
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.19.2
